+++ packages/suite/src/actions/suite/analyticsActions.ts
@@ -10,8 +10,8 @@ import { ANALYTICS } from '@suite-actions/constants';
 import { getTrackingRandomId } from '@suite-utils/random';
 import { allowSentryReport, setSentryUser } from '@suite-utils/sentry';
 import { AnalyticsState } from '@suite-reducers/analyticsReducer';
-import { isDev } from '@suite-utils/build';
 import { getEnvironment } from '@suite-utils/env';
+import { hasUserAllowedTracking } from '@suite-utils/analytics';
 
 import type { Dispatch } from '@suite-types';
 
@@ -29,7 +29,7 @@ export type AnalyticsAction =
       };
 
 export const enable = () => (dispatch: Dispatch) => {
-    analytics.report({ type: EventType.AnalyticsEnable });
+    analytics.report({ type: EventType.SettingsAnalytics, payload: { value: true } });
     allowSentryReport(true);
 
     dispatch({
@@ -38,7 +38,7 @@ export const enable = () => (dispatch: Dispatch) => {
 };
 
 export const disable = () => (dispatch: Dispatch) => {
-    analytics.report({ type: EventType.AnalyticsDispose }, true);
+    analytics.report({ type: EventType.SettingsAnalytics, payload: { value: false } }, true);
     allowSentryReport(false);
 
     dispatch({
@@ -56,20 +56,14 @@ export const init = (state: AnalyticsState) => (dispatch: Dispatch) => {
     // if instanceId does not exist yet (was not loaded from storage), create a new one
     const instanceId = state.instanceId || getTrackingRandomId();
     const sessionId = getTrackingRandomId();
-
-    // if user made choice, keep it, otherwise set it to true by default just to prefill the confirmation toggle
-    const isConfirmed = !!state.confirmed;
-    const isEnabled = isConfirmed ? !!state.enabled : true;
-
-    // allow tracking only if user already confirmed data collection
-    const userAllowedTracking = isConfirmed && isEnabled;
+    const userAllowedTracking = hasUserAllowedTracking(state.enabled, state.confirmed);
 
     analytics.init(userAllowedTracking, {
         instanceId,
         sessionId,
         environment: getEnvironment(),
         commitId: process.env.COMMITHASH || '',
-        isDev,
+        isDev: !process.env.CODESIGN_BUILD,
         callbacks: {
             onEnable: () => dispatch(enable()),
             onDisable: () => dispatch(disable()),
@@ -84,8 +78,9 @@ export const init = (state: AnalyticsState) => (dispatch: Dispatch) => {
         payload: {
             instanceId,
             sessionId,
-            enabled: isEnabled,
-            confirmed: isConfirmed,
+            // if user made choice, keep it, otherwise set it to true by default just to prefill the confirmation toggle
+            enabled: state.confirmed ? !!state.enabled : true,
+            confirmed: !!state.confirmed,
         },
     });
 };