+++ packages/suite/src/components/suite/NavigationBar/components/NavigationActions/index.tsx
@@ -1,6 +1,5 @@
 import React, { useMemo } from 'react';
 import styled from 'styled-components';
-import { analytics, EventType } from '@trezor/suite-analytics';
 
 import * as walletSettingsActions from '@settings-actions/walletSettingsActions';
 import * as suiteActions from '@suite-actions/suiteActions';
@@ -76,18 +75,8 @@ export const NavigationActions: React.FC<NavigationActionsProps> = ({
 
     const WrapperComponent = isMobileLayout ? MobileWrapper : Wrapper;
 
-    // used only in mobile layout
-    const gotoWithReport = (routeName: Route['name']) => {
-        if (routeName === 'notifications-index') {
-            analytics.report({ type: EventType.MenuGotoNotificationsIndex });
-        } else if (routeName === 'settings-index') {
-            analytics.report({ type: EventType.MenuGotoSettingsIndex });
-        }
-        goto(routeName);
-    };
-
     const action = (route: Route['name']) => {
-        gotoWithReport(route);
+        goto(route);
         if (closeMainNavigation) closeMainNavigation();
     };
 
@@ -110,15 +99,7 @@ export const NavigationActions: React.FC<NavigationActionsProps> = ({
     return (
         <WrapperComponent>
             <ActionItem
-                onClick={() => {
-                    analytics.report({
-                        type: EventType.MenuToggleDiscreet,
-                        payload: {
-                            value: !discreetMode,
-                        },
-                    });
-                    setDiscreetMode(!discreetMode);
-                }}
+                onClick={() => setDiscreetMode(!discreetMode)}
                 isActive={discreetMode}
                 label={<Translation id="TR_DISCREET" />}
                 icon={discreetMode ? 'HIDE' : 'SHOW'}
@@ -157,10 +138,7 @@ export const NavigationActions: React.FC<NavigationActionsProps> = ({
             {allowPrerelease &&
                 (isMobileLayout ? (
                     <ActionItem
-                        onClick={() => {
-                            analytics.report({ type: EventType.MenuGotoEarlyAccess });
-                            action('settings-index');
-                        }}
+                        onClick={() => action('settings-index')}
                         label={<Translation id="TR_EARLY_ACCESS_MENU" />}
                         icon="EXPERIMENTAL_FEATURES"
                         isMobileLayout