+++ packages/suite/src/utils/suite/analytics.ts
@@ -13,8 +13,25 @@ import {
     getPlatformLanguages,
 } from '@suite-utils/env';
 import { getCustomBackends } from '@suite-utils/backend';
+import { AccountTransactionBaseAnchor } from '@suite-constants/anchors';
 
+import type { AnalyticsState } from '@suite-reducers/analyticsReducer';
 import type { UpdateInfo } from '@trezor/suite-desktop-api';
+
+// redact transaction id from account transaction anchor
+export const redactTransactionIdFromAnchor = (anchor?: string) => {
+    if (!anchor) {
+        return undefined;
+    }
+
+    return anchor.startsWith(AccountTransactionBaseAnchor) ? AccountTransactionBaseAnchor : anchor;
+};
+
+// allow tracking only if user already confirmed data collection
+export const hasUserAllowedTracking = (
+    enabled: AnalyticsState['enabled'],
+    confirmed: AnalyticsState['confirmed'],
+) => !!confirmed && !!enabled;
 
 export const getSuiteReadyPayload = (state: AppState) => ({
     language: state.suite.settings.language,